Fix cross-platform build issues
- Fix Boost filesystem API usage (copy_option -> copy_options)
- Fix logical operator precedence in file_manager.cpp
- Fix std::min template ambiguity in key_generator.cpp
- Add vcpkg.json for better Windows dependency management
- Improve Boost detection for vcpkg environments
- Update CI workflow to use vcpkg.json

+++ src/utils/key_generator.cpp
@@ -137,7 +137,7 @@ KeyGenerator::KeyStrength KeyGenerator::evaluatePasswordStrength(const std::stri
     strength.score = 0;
     
     // Length scoring
-    int lengthScore = std::min(password.length() * 2, 25UL);
+    int lengthScore = std::min(static_cast<size_t>(password.length() * 2), 25UL);
     strength.score += lengthScore;
     
     // Character variety scoring
